refactor: using onboarding-sdk common instead of selc-commons
