+++ apps/institution-ms/connector-api/src/main/java/it/pagopa/selfcare/mscore/model/onboarding/OnboardingRequest.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.mscore.model.onboarding;
 
-import it.pagopa.selfcare.commons.base.utils.InstitutionType;
+import it.pagopa.selfcare.onboarding.common.InstitutionType;
 import it.pagopa.selfcare.mscore.constant.TokenType;
 import it.pagopa.selfcare.mscore.exception.InvalidRequestException;
 import it.pagopa.selfcare.mscore.model.institution.Billing;
