+++ apps/institution-ms/connector/dao/src/test/java/it/pagopa/selfcare/mscore/connector/dao/utils/TestUtils.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.mscore.connector.dao.utils;
 
-import it.pagopa.selfcare.commons.base.utils.InstitutionType;
+import it.pagopa.selfcare.onboarding.common.InstitutionType;
 import it.pagopa.selfcare.mscore.connector.dao.model.InstitutionEntity;
 import it.pagopa.selfcare.mscore.connector.dao.model.inner.BillingEntity;
 import it.pagopa.selfcare.mscore.connector.dao.model.inner.DataProtectionOfficerEntity;
