+++ apps/institution-ms/core/src/test/java/it/pagopa/selfcare/mscore/core/util/TestUtils.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.mscore.core.util;
 
-import it.pagopa.selfcare.commons.base.utils.InstitutionType;
+import it.pagopa.selfcare.onboarding.common.InstitutionType;
 import it.pagopa.selfcare.mscore.constant.Origin;
 import it.pagopa.selfcare.mscore.model.institution.*;
 import it.pagopa.selfcare.mscore.model.onboarding.Contract;
