+++ apps/institution-ms/web/src/main/java/it/pagopa/selfcare/mscore/web/model/onboarding/OnboardedInstitutionResponse.java
@@ -2,7 +2,7 @@ package it.pagopa.selfcare.mscore.web.model.onboarding;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import it.pagopa.selfcare.commons.base.security.PartyRole;
-import it.pagopa.selfcare.commons.base.utils.InstitutionType;
+import it.pagopa.selfcare.onboarding.common.InstitutionType;
 import it.pagopa.selfcare.mscore.model.institution.Billing;
 import it.pagopa.selfcare.mscore.web.model.institution.AttributesResponse;
 import it.pagopa.selfcare.mscore.web.model.institution.DataProtectionOfficerResponse;
