+++ apps/institution-ms/web/src/main/java/it/pagopa/selfcare/mscore/web/model/token/InstitutionToNotifyResponse.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.mscore.web.model.token;
 
-import it.pagopa.selfcare.commons.base.utils.InstitutionType;
+import it.pagopa.selfcare.onboarding.common.InstitutionType;
 import it.pagopa.selfcare.mscore.model.RootParent;
 import it.pagopa.selfcare.mscore.model.institution.PaymentServiceProvider;
 import lombok.AllArgsConstructor;
