+++ apps/institution-ms/web/src/test/java/it/pagopa/selfcare/mscore/web/TestUtils.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.mscore.web;
 
-import it.pagopa.selfcare.commons.base.utils.InstitutionType;
+import it.pagopa.selfcare.onboarding.common.InstitutionType;
 import it.pagopa.selfcare.mscore.constant.Origin;
 import it.pagopa.selfcare.mscore.model.institution.Billing;
 import it.pagopa.selfcare.mscore.model.institution.DataProtectionOfficer;
