+++ apps/institution-ms/web/src/test/java/it/pagopa/selfcare/mscore/web/controller/ExternalControllerTest.java
@@ -1,6 +1,6 @@
 package it.pagopa.selfcare.mscore.web.controller;
 
-import it.pagopa.selfcare.commons.base.utils.InstitutionType;
+import it.pagopa.selfcare.onboarding.common.InstitutionType;
 import it.pagopa.selfcare.mscore.constant.RelationshipState;
 import it.pagopa.selfcare.mscore.core.ExternalService;
 import it.pagopa.selfcare.mscore.model.institution.*;
